moved actions to new actions.js file
